WIP: initial screen reader interactions
By default a screen reader will intercept key press events to navigate a
virtual DOM the screen reader manages. For Strype, we need those
key presses to go to the Vue application and interact with our app.

This is the first of many modifications towards making Strype usable with
a screen reader.

Signed-off-by: Joshua Lock <joshua.lock@kcl.ac.uk>